+++ NEWS.md
@@ -1,3 +1,9 @@
+# osrmr 0.1.29
+
+* Set examples using OSRMR webserver to "dontrun" because of check errors due to flaky web server
+* Bugfix in viaroute when using instructions
+* Updated documentation
+
 # osrmr 0.1.28
 
 * Added a `NEWS.md` file to track changes to the package.